+++ app/src/main/java/com/example/newsapp/model/ContactInformationType.kt
@@ -1,0 +1,7 @@
+package com.example.newsapp.model
+
+enum class ContactInformationType {
+    PHONE,
+    EMAIL,
+    LINKEDIN
+}
