+++ app/src/main/java/com/example/newsapp/model/data/local/daos/NewsDao.kt
@@ -1,10 +1,10 @@
-package com.example.newsapp.data.local.daos
+package com.example.newsapp.model.data.local.daos
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.example.newsapp.data.local.entities.NewsEntity
+import com.example.newsapp.model.data.local.entities.NewsEntity
 
 @Dao
 interface NewsDao {
