+++ app/src/main/java/com/example/newsapp/view/NewsListScreen.kt
@@ -48,7 +48,7 @@ import androidx.navigation.compose.rememberNavController
 import coil.compose.AsyncImage
 import com.example.newsapp.R
 import com.example.newsapp.Routes
-import com.example.newsapp.data.local.entities.NewsEntity
+import com.example.newsapp.model.data.local.entities.NewsEntity
 import com.example.newsapp.ui.theme.LightRed
 import com.example.newsapp.ui.theme.NewsAppTheme
 import com.example.newsapp.ui.theme.NoInternetBannerBtnStyle
