+++ app/src/main/java/com/example/newsapp/viewmodel/NewsListScreenViewModel.kt
@@ -4,8 +4,8 @@ import android.util.Log
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.newsapp.data.local.entities.NewsEntity
-import com.example.newsapp.data.local.repositories.NewsRepositoryImpl
+import com.example.newsapp.model.data.local.entities.NewsEntity
+import com.example.newsapp.model.data.local.repositories.NewsRepositoryImpl
 import com.example.newsapp.util.Constants.DEFAULT_LOADING_DURATION_IN_MILLIS
 import com.example.newsapp.util.Constants.PAGE_SIZE
 import com.example.newsapp.util.Resource
@@ -49,10 +49,13 @@ class NewsListScreenViewModel @Inject constructor(
 
     fun loadNewsListPaginated() {
         viewModelScope.launch {
+            // If the news are loaded initially, display a circular progress indicator and give a little time to the user
+            // to see the loading process
             if (newsList.value.isEmpty()) {
                 isLoadingInitialNews.value = true
                 delay(DEFAULT_LOADING_DURATION_IN_MILLIS)
             }
+
             val newsResponseWithTotalResults = repository.getNewsResponseWithTotalResults(page = currentPage)
             val newsEntitiesList = newsResponseWithTotalResults.data?.first
             val totalResults = newsResponseWithTotalResults.data?.second
@@ -62,9 +65,10 @@ class NewsListScreenViewModel @Inject constructor(
                     paginationEndReached.value = currentPage * PAGE_SIZE >= (totalResults ?: PAGE_SIZE)
                     currentPage++
                     loadError.value = false
+                    // If the news are loaded initially or are being refreshed and the response is successful,
+                    // the database is updated, so the old data are deleted
                     if (isLoadingInitialNews.value || refreshing.value) {
                         if (newsEntitiesList != null) {
-                            println("NewsDB: deleted")
                             repository.deleteDb()
                         }
                         isLoadingInitialNews.value = false
@@ -77,7 +81,8 @@ class NewsListScreenViewModel @Inject constructor(
                 }
 
                 is Resource.Error -> {
-                    // If the news cannot be loaded, give a little time to the user to see the loading bar before the error message is displayed
+                    // If the news cannot be loaded, give a little time to the user to see the circular progress indicator
+                    // before the error message is displayed
                     delay(DEFAULT_LOADING_DURATION_IN_MILLIS)
                     loadError.value = true
                     Log.e("Failed to load News Articles.", "Reason: ${ newsResponseWithTotalResults.message }")
