+++ app/src/main/java/com/example/newsapp/viewmodel/WebViewScreenViewModel.kt
@@ -4,9 +4,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
@@ -17,20 +15,18 @@ class WebViewScreenViewModel @Inject constructor(
         private set
 
     fun onEvent(event: WebViewScreenEvent) {
-        viewModelScope.launch {
-            state = when (event) {
-                WebViewScreenEvent.OnLoadingError -> {
-                    state.copy(
-                        isLoading = false,
-                        loadError = true
-                    )
-                }
+        state = when (event) {
+            WebViewScreenEvent.OnLoadingError -> {
+                state.copy(
+                    isLoading = false,
+                    loadError = true
+                )
+            }
 
-                is WebViewScreenEvent.OnLoadingInProgress -> {
-                    state.copy(
-                        isLoading = event.isLoading
-                    )
-                }
+            is WebViewScreenEvent.OnLoadingInProgress -> {
+                state.copy(
+                    isLoading = event.isLoading
+                )
             }
         }
     }
